klip-selector/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0
apscheduler>=3.10,<4.0
redis>=5.0.0
paapi5-python-sdk==1.0.0
requests>=2.31.0
